one more minor

+++ scripts/Embedded/Examples/FluidStructure/Burman-Stokes/Generate_visualisations.jl
@@ -233,8 +233,10 @@ function main(ranks)
     writevtk(Ω.Ωs,files_path*"Omega_s_$it",cellfields=["uh"=>uh,"ph"=>ph,"dh"=>dh])
     toc!(t,"Write")
 
-    i_am_main(ranks) && run(`tar -czf $files_path/data_$it.tar.gz $files_path/Omega_s_$it $files_path/Omega_f_$it $files_path/Omega_act_$it`) &&
+    if i_am_main(ranks)
+      run(`tar -czf $files_path/data_$it.tar.gz $files_path/Omega_s_$it $files_path/Omega_f_$it $files_path/Omega_act_$it`)
       run(`rm -r $files_path/Omega_s_$it $files_path/Omega_f_$it $files_path/Omega_act_$it`)
+    end
   end
 end
